test: extract submission handler white-box suites
