+++ crates/agent-engine/src/runtime/submission_handlers/tests.rs
@@ -1,0 +1,3 @@
+include!("tests/confirmation_and_mount_support.inc.rs");
+include!("tests/mount_and_user_input_contract.inc.rs");
+include!("tests/runtime_ops_and_replay_contract.inc.rs");
